It is working. (Without elo and with bugs)
It's buggy at this state. But should work if you handle it correctly.

Known bugs: When people can't connect and it says "waiting for endpoint" on discord.

The bot adds itself to the queue when starting it up. (temporary fix by just typing "rs" in the console)

People cant unready,

name is undefined on the player object inside the queue.

As you can see, there is some things that have to be fixed before we can test it.

What I've done in this commit:
fixed a matching queue or a game object that gets saved to a json file that stores some basic information about the match. This information contains:
{
  "index": [
    {
      "key": "match",
      "val": {
        "matchid": 90,
        "hasStarted": false,
        "allPlayersConfirmed": true,
        "team1": [
          {
            "id": "223318872323784704",
            "confirmed": true,
            "confirmable": true
          }
        ],
        "team2": []
      },
      "life": 1560368564463,
      "tags": []
    }
  ]
}

In future commits I will fix the remaining big bugs and fix so that the bot moves players to specific rooms with their team inside. (and of course elo)